tecnostar/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.1
dnspython==2.6.1
Markdown==3.6
pillow==10.3.0
pymongo==4.7.0